Move lineRangesToCharRanges
