+++ palantir-java-format-utils/build.gradle
@@ -2,4 +2,11 @@ apply from: rootProject.file('gradle/publish-jar.gradle')
 
 dependencies {
   implementation project(':palantir-java-format-spi')
+
+    testImplementation 'com.google.guava:guava-testlib'
+    testImplementation 'com.google.testing.compile:compile-testing'
+    testImplementation 'com.google.truth:truth'
+    testImplementation 'org.assertj:assertj-core'
+    testImplementation 'org.junit.jupiter:junit-jupiter-migrationsupport'
+    testImplementation 'org.junit.jupiter:junit-jupiter'
 }
