+++ palantir-java-format-utils/src/main/java/com/palantir/javaformat/Newlines.java
